0.2

+++ edgelord/security_data.py
@@ -29,10 +29,9 @@ class security_data:
         self.parent = parent
 
         # Check caching
-        if self.caching and 'EDGELORD_CACHE_DIR' in os.environ:
+        if 'EDGELORD_CACHE_DIR' in os.environ:
             self.cache_directory = str(os.environ['EDGELORD_CACHE_DIR']).replace("~", os.environ['HOME'])
             if os.path.exists(self.cache_directory):
-
                 days_before_renew = 2
                 if 'EDGELORD_CACHE_DAYS' in os.environ:
                     day_before_renew = int(os.environ['EDGELORD_CACHE_DAYS'])
@@ -42,15 +41,14 @@ class security_data:
                 hours = (minutes / 60)
                 days = (hours / 24)
 
-                if days_before_renew >= days:
+                if days > 1 and days_before_renew < days:
                     os.rmdir(self.cache_directory)
+            else:
+                os.mkdir(self.cache_directory)
 
-            #if not os.path.exists(self.cache_directory):
-            #    os.mkdir(self.cache_directory)
-
-    def __del__(self):
-        if self.caching and self.cache_directory is not None:
-            self.frame().to_json("%s/%s.json" % (self.cache_directory, self.parent.ticker))
+#    def __del__(self):
+#        if self.caching and self.cache_directory is not None:
+#            self.frame().to_json("%s/%s.json" % (self.cache_directory, self.parent.ticker))
 
     def __str__(self):
         return str(self.frame())
@@ -74,6 +72,10 @@ class security_data:
 
         return pandas.DataFrame(self.data, columns = self.columns).sort_values(by=['date'])
 
+    def from_cache(self):
+        print("...")
+        return 1
+
     def from_csv(self, file):
         """Build a security_data object from the given CSV file"""
         
@@ -95,43 +97,51 @@ class security_data:
     def from_rest(self, ticker = None, days = 260):
         """Download data from Tiingo"""
 
-        client = TiingoClient()
-        
-        data = client.get_ticker_price(ticker, startDate=last_business_day(days), endDate=last_business_day())
 
-        result = []
+        self.cache_filename = "%s/%s.json" % (self.cache_directory, ticker)
+        if os.path.exists(self.cache_filename):
+            buffer = self.from_json(self.cache_filename)
+            print("Loading from cache")
+        else:
+            client = TiingoClient()
+            data = client.get_ticker_price(ticker, startDate=last_business_day(days), endDate=last_business_day())
 
-        column_list = None
-        set_columns = False
-        for row in data:
-            buffer = {}
-            
-            if column_list is None and not set_columns:
-                column_list = []
+            result = []
 
-            for column in row:
-                # Loop through each rows column list
-                if ("adj" in column) or (column == 'date'):
-                    # Only grab adjusted data (to account for any splits)
+            column_list = None
+            set_columns = False
+            for row in data:
+                buffer = {}
+                
+                if column_list is None and not set_columns:
+                    column_list = []
 
-                    column_name = column.replace("adj", "").lower()
+                for column in row:
+                    # Loop through each rows column list
+                    if ("adj" in column) or (column == 'date'):
+                        # Only grab adjusted data (to account for any splits)
 
-                    if not set_columns:
-                        column_list.append(column.replace("adj", "").lower())
+                        column_name = column.replace("adj", "").lower()
 
-                    if column == 'date':
-                        buffer[column_name] = row[column].split("T")[0]
-                    else:
-                        buffer[column_name] = row[column]
+                        if not set_columns:
+                            column_list.append(column.replace("adj", "").lower())
 
-            if not set_columns and column_list is not None:
-                # Only set the column list once
-                set_columns = True
+                        if column == 'date':
+                            buffer[column_name] = row[column].split("T")[0]
+                        else:
+                            buffer[column_name] = row[column]
 
-            result.append(buffer)
+                if not set_columns and column_list is not None:
+                    # Only set the column list once
+                    set_columns = True
 
-        data = pandas.DataFrame(result, columns=column_list)
-        data = data.sort_values(by=['date'])
+                result.append(buffer)
 
-        return self.build(data)
+            data = pandas.DataFrame(result, columns=column_list)
+            data = data.sort_values(by=['date'])
+
+            data.to_json("%s/%s.json" % (self.cache_directory, ticker))
+            buffer = self.build(data)
+
+        return buffer
